requirements_common.txt
# The necessary dependencies common to both local and heroku installs

# Mental note (cmd to check for updates): pip list --outdated --format=columns
# Mental note (npm equivalent command):   npm outdated

# Core

Django==1.10.6
django-appconf==1.0.2                   # Helper for handling app configs
django-dynamic-preferences==1.1.0       # Settings management
django-extensions==1.7.7                # For the generate secret command
django-ipware==1.1.6                    # IP Address logging
django-jet==1.0.5                       # Admin Backend
django-gfklookupwidget==1.0.5           # Replaces object_id field with a search link

# Database

psycopg2==2.7.1                         # For Django to talk to postgres
sqlparse==0.2.3                         # Parsing SQL statements

# Pipeline

django-gulp==2.5.0                      # Run gulp tasks with dj r/collectstatic
whitenoise==3.3.0                       # Static file serving
waitress==1.0.2                         # For web serving

# Development Aids

dj-cmd==1.0                             # Provides the dj command alias
django-debug-toolbar==1.7               # Debug Toolbar
six==1.10                               # Python 2/3 compatibility library
flake8==3.3.0                           # For testing/linting
flake8-import-order==0.12               # Flake plugin for import order
flake8-quotes==0.9.0                    # Flake plugin for quotes
pep8-naming==0.4.1                      # Flake plugin for naming conventions
